move MockContext to test_utils
Will be used by the II term benches.
